class work

+++ src/main/java/in/ashokit/dto/DashboardResponseDTO.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class DashboardResponseDTO {
 	
-	private Long totalEnqCnt;
-	private Long openEnqCnt;
-	private Long enrolledEnqCnt;
-	private Long lostEnqCnt;
+	private Integer totalEnqCnt;
+	private Integer openEnqCnt;
+	private Integer enrolledEnqCnt;
+	private Integer lostEnqCnt;
 
 }
